refactor: move all malloc v1 code out of src
