+++ malloc_v1/heap.c
@@ -1,5 +1,5 @@
 
-#include "../../inc/main.h" 
+#include "./malloc_v1.h" 
 
 void create_heap(t_heap** heap, size_t size) {
   *heap = (t_heap*)mmap(0, size, PROT_READ | PROT_WRITE, MAP_ANON | MAP_PRIVATE, -1, 0);
@@ -8,7 +8,6 @@ void create_heap(t_heap** heap, size_t size) {
   (*heap)->total_size = size;
   (*heap)->free_size = size;
   (*heap)->block_count = 0;
-  (*heap)->cache_table = NULL;
 }
 
 void destroy_heap(t_heap* heap) {
