+++ malloc_v1/my_realloc.c
@@ -1,4 +1,4 @@
-#include "../../inc/main.h" 
+#include "./malloc_v1.h" 
 
 void* my_realloc(void* ptr, size_t size) {
     // to compile the code with unused params, we need to use the UNUSED macro
