+++ src/utils.c
@@ -2,7 +2,6 @@
 
 size_t min_chunk_size = 0;
 size_t pointer_size = 0;
-t_heap* global_heap = NULL;
 
 void log_info(const char* message) {
     printf("\n=====%s=====\n", message);
